extract PublicFieldsFinder to reuse
